# agent/openapi.yaml
openapi: 3.1.0
info:
  title: Investment Portfolio API
  version: 0.1.1
  description: API for searching funds, computing summary stats, portfolio ranking, and RAG retrieval.

servers:
  - url: https://investassistant.onrender.com

tags:
  - name: Health
    description: Health checks
  - name: Funds
    description: Fund search and analytics
  - name: Portfolio
    description: Portfolio construction and optimization
  - name: RAG
    description: Retrieval endpoints for RAG workflows

paths:
  /healthz:
    get:
      tags: [Health]
      operationId: healthz
      summary: Health check
      description: Returns service status.
      x-openai-isConsequential: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /funds/search:
    post:
      tags: [Funds]
      operationId: searchFunds
      summary: Search funds (table query)
      description: |
        Query a table (typically fund_data) with optional filters, selected columns, ordering, and limit. filters is a list of {column, op, value}. Returns {"rows":[...]}.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundSearchRequest"
            examples:
              minimal_working:
                value:
                  table: fund_data
                  query: "S&P"
                  limit: 5
              with_filters_and_order:
                value:
                  table: fund_data
                  columns: ["id", "fund_name", "expense_ratio", "ytd_return_percentage"]
                  filters:
                    - column: expense_ratio
                      op: "<="
                      value: 0.15
                    - column: ytd_return_percentage
                      op: ">="
                      value: 0
                  limit: 10
                  order_by: "-ytd_return_percentage"
      responses:
        "200":
          description: Rows matching query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RowsOnlyResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /funds/summary:
    post:
      tags: [Funds]
      operationId: summarizeFunds
      summary: Summary statistics (min/max/avg) for selected columns
      description: Computes MIN/MAX/AVG for requested columns, with optional filters.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundSummaryRequest"
            examples:
              example:
                value:
                  table: fund_data
                  columns: ["expense_ratio", "ytd_return_percentage", "3_year_volatility"]
                  filters:
                    - column: expense_ratio
                      op: "<="
                      value: 0.15
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundSummaryResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /portfolio/lowest-expense:
    post:
      tags: [Portfolio]
      operationId: lowestExpensePortfolio
      summary: Return N rows with lowest expense column
      description: Returns lowest values for expense_column with optional filters.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LowestExpenseRequest"
            examples:
              example:
                value:
                  table: fund_data
                  expense_column: expense_ratio
                  count: 10
                  filters:
                    - column: ytd_return_percentage
                      op: ">="
                      value: 0
      responses:
        "200":
          description: Portfolio rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RowsWithCountResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /portfolio/best-return:
    post:
      tags: [Portfolio]
      operationId: bestReturnPortfolio
      summary: Return N rows with best return column
      description: Returns highest values for return_column with optional filters.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BestReturnRequest"
            examples:
              example:
                value:
                  table: fund_data
                  return_column: ytd_return_percentage
                  count: 10
                  filters:
                    - column: expense_ratio
                      op: "<="
                      value: 0.2
      responses:
        "200":
          description: Portfolio rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RowsWithCountResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /portfolio/optimize:
    post:
      tags: [Portfolio]
      operationId: optimizePortfolio
      summary: Optimize (rank) using weighted objective
      description: Ranks rows using objective (maximize/minimize) and optional weights and filters.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OptimizeRequest"
            examples:
              example:
                value:
                  table: fund_data
                  count: 10
                  objective:
                    maximize: ["ytd_return_percentage"]
                    minimize: ["expense_ratio", "3_year_volatility"]
                  weights:
                    ytd_return_percentage: 0.7
                    expense_ratio: 0.2
                    3_year_volatility: 0.1
                  filters:
                    - column: expense_ratio
                      op: "<="
                      value: 0.25
      responses:
        "200":
          description: Portfolio rows with score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RowsWithCountResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /rag/search:
    post:
      tags: [RAG]
      operationId: ragSearch
      summary: Keyword RAG search
      description: Keyword retrieval over text_columns; returns ranked rows.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagSearchRequest"
            examples:
              example:
                value:
                  query: "low expense ratio US equity"
                  table: fund_data
                  columns: ["id", "fund_name", "expense_ratio"]
                  text_columns: ["fund_name", "morningstart_risk"]
                  limit: 25
      responses:
        "200":
          description: RAG keyword results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagRowsResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

  /rag/semantic:
    post:
      tags: [RAG]
      operationId: ragSemanticSearch
      summary: Semantic RAG search
      description: Vector similarity retrieval using embedding and embeddings_table.
      x-openai-isConsequential: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RagSemanticRequest"
            examples:
              example:
                value:
                  embedding: [0.01, -0.02, 0.03]
                  table: fund_data
                  embeddings_table: fund_embeddings
                  columns: ["id", "fund_name", "expense_ratio"]
                  limit: 10
      responses:
        "200":
          description: RAG semantic results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RagRowsSemanticResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FastApiError"
        "500":
          description: Server error
          content:
            text/plain:
              schema:
                type: string

components:
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      properties:
        status:
          type: string
      required: [status]

    Filter:
      type: object
      additionalProperties: false
      properties:
        column:
          type: string
          description: Column name in the selected table
        op:
          type: string
          enum: ["=", "!=", ">", ">=", "<", "<=", "in", "like"]
        value:
          description: Value for the filter; for op=in must be an array.
          oneOf:
            - type: string
            - type: number
            - type: boolean
            - type: array
              items:
                oneOf:
                  - type: string
                  - type: number
                  - type: boolean
      required: [column, op, value]

    FundSearchRequest:
      type: object
      additionalProperties: false
      properties:
        table:
          type: string
          default: fund_data
        columns:
          description: Selected columns; omit to return all. Use ["*"] for all.
          anyOf:
            - type: "null"
            - type: array
              items: { type: string }
        filters:
          type: array
          items:
            $ref: "#/components/schemas/Filter"
          default: []
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
        order_by:
          description: Column name to order by; prefix with '-' for DESC.
          anyOf:
            - type: "null"
            - type: string
        query:
          description: Optional search string (some deployments may ignore this).
          anyOf:
            - type: "null"
            - type: string
      required: [table]

    FundSummaryRequest:
      type: object
      additionalProperties: false
      properties:
        table:
          type: string
          default: fund_data
        columns:
          type: array
          minItems: 1
          items: { type: string }
        filters:
          type: array
          items:
            $ref: "#/components/schemas/Filter"
          default: []
      required: [table, columns]

    LowestExpenseRequest:
      type: object
      additionalProperties: false
      properties:
        table:
          type: string
          default: fund_data
        expense_column:
          type: string
        count:
          type: integer
          minimum: 1
          maximum: 1000
          default: 10
        filters:
          type: array
          items:
            $ref: "#/components/schemas/Filter"
          default: []
      required: [table, expense_column]

    BestReturnRequest:
      type: object
      additionalProperties: false
      properties:
        table:
          type: string
          default: fund_data
        return_column:
          type: string
        count:
          type: integer
          minimum: 1
          maximum: 1000
          default: 10
        filters:
          type: array
          items:
            $ref: "#/components/schemas/Filter"
          default: []
      required: [table, return_column]

    Objective:
      type: object
      additionalProperties: false
      properties:
        maximize:
          type: array
          items: { type: string }
          default: []
        minimize:
          type: array
          items: { type: string }
          default: []
      required: [maximize, minimize]

    OptimizeRequest:
      type: object
      additionalProperties: false
      properties:
        table:
          type: string
          default: fund_data
        count:
          type: integer
          minimum: 1
          maximum: 1000
          default: 10
        objective:
          $ref: "#/components/schemas/Objective"
        weights:
          type: object
          additionalProperties:
            type: number
          default: {}
        filters:
          type: array
          items:
            $ref: "#/components/schemas/Filter"
          default: []
      required: [table, count, objective]

    RagSearchRequest:
      type: object
      additionalProperties: false
      properties:
        query:
          type: string
        table:
          type: string
          default: fund_data
        columns:
          anyOf:
            - type: "null"
            - type: array
              items: { type: string }
        text_columns:
          anyOf:
            - type: "null"
            - type: array
              items: { type: string }
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
      required: [query]

    RagSemanticRequest:
      type: object
      additionalProperties: false
      properties:
        embedding:
          type: array
          minItems: 1
          items: { type: number }
        table:
          type: string
          default: fund_data
        embeddings_table:
          type: string
          default: fund_embeddings
        columns:
          anyOf:
            - type: "null"
            - type: array
              items: { type: string }
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
      required: [embedding]

    AnyRow:
      type: object
      properties: {}
      additionalProperties: true
      description: Arbitrary row object; fields depend on requested columns.

    RowsOnlyResponse:
      type: object
      additionalProperties: false
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AnyRow"
      required: [rows]

    RowsWithCountResponse:
      type: object
      additionalProperties: false
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AnyRow"
        count:
          type: integer
      required: [rows, count]

    FundSummaryMetric:
      type: object
      additionalProperties: false
      properties:
        min:
          type: ["number", "null"]
        max:
          type: ["number", "null"]
        avg:
          type: ["number", "null"]
      required: [min, max, avg]

    FundSummaryResponse:
      type: object
      additionalProperties: false
      properties:
        summary:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/FundSummaryMetric"
      required: [summary]

    RagRowsResponse:
      type: object
      additionalProperties: false
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AnyRow"
        count:
          type: integer
      required: [rows, count]

    RagRowsSemanticResponse:
      type: object
      additionalProperties: false
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/AnyRow"
        count:
          type: integer
      required: [rows, count]

    FastApiError:
      type: object
      additionalProperties: true
      properties:
        detail:
          additionalProperties: true